chore: move useUpdateOnlyEffectHook into the parts folder #1687
