+++ ui/src/table.tsx
@@ -22,7 +22,7 @@ import { MarkdownTableCellType, XMarkdownTableCellType } from './markdown_table_
 import { ProgressTableCellType, XProgressTableCellType } from "./progress_table_cell_type"
 import { TagTableCellType, XTagTableCellType } from "./tag_table_cell_type"
 import { border, cssVar, important, margin, rem } from './theme'
-import useUpdateOnlyEffect from './useUpdateOnlyEffectHook'
+import useUpdateOnlyEffect from './parts/useUpdateOnlyEffectHook'
 import { wave } from './ui'
 
 /** Configure table pagination. Use as `pagination` parameter to `ui.table()` */
